fix pr review

+++ src/Application/RolePermissions/Delete/DeleteRolePermissionCommandHandler.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Application.Abstractions.Data;
+﻿using Application.Abstractions.Data;
 using Application.Abstractions.Messaging;
 using Domain.RolePermissions;
 using Microsoft.EntityFrameworkCore;
@@ -22,8 +19,11 @@ public sealed class DeleteRolePermissionCommandHandler
     public async Task<Result> Handle(DeleteRolePermissionCommand command, CancellationToken cancellationToken)
     {
         RolePermission? rolePermission = await _context.RolePermissions
-            .FindAsync(new object[] { command.RoleId, command.PermissionId }, cancellationToken) ?? throw new InvalidOperationException("Role permission not found.");
-
+            .FindAsync(new object[] { command.RoleId, command.PermissionId }, cancellationToken);
+        if(rolePermission is null)
+        {
+            return Result.Failure<Guid>("role Permission  not found.");
+        }
         _context.RolePermissions.Remove(rolePermission);
         await _context.SaveChangesAsync(cancellationToken);
